Implement initial placement version 1 (Sequential, Center-first, Interdigital-based Center-first)
Change aspect format from int to double

+++ header/InitialPlacement.h
@@ -8,14 +8,14 @@ private:
     std::string circuitType;
     int group = 0;
     int nfin = 0;
-	int aspect = 0;
+	double aspect = 0.0;
     std::vector<int> deviceNumList;
 
     std::vector<TableManager> tableManager;
 	void GenerateInitialPlacementTables();
 
 public:
-    InItialPlacenent(std::string circuitType, int group, int nfin, int aspect, std::vector<int>& deviceNumList);
+    InItialPlacenent(std::string circuitType, int group, int nfin, double aspect, std::vector<int>& deviceNumList);
 
     std::vector<TableManager> GetResultTable() { return tableManager; }
 };
